Move stack frame and instruction into files.
